requirements.txt
lark-parser>=0.12.0
colorama>=0.4.6
